Refactored edge routing
